flask_app/requirements.txt
absl-py==0.15.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
clang==5.0
click==8.0.3
cycler==0.11.0
fastapi==0.73.0
flatbuffers==1.12
fonttools==4.29.0
gast==0.4.0
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.13.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.10.1
install==1.3.5
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.15.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
uvicorn==0.17.1
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.7.0
